Holi boli

+++ src/negocio/entidades/ProductoProveedor.java
@@ -8,8 +8,8 @@ public class ProductoProveedor extends Producto {
 
     private Proveedor proveedorDelProducto_;
 
-    public ProductoProveedor(int idProducto, String nombre, float precio, 
-            float costo, Proveedor proveedor) {
+    public ProductoProveedor(int idProducto, String nombre, double precio, 
+            double costo, Proveedor proveedor) {
         super(idProducto, nombre, precio, costo);
         this.proveedorDelProducto_ = proveedor;
     }
